feat(assembleExportDiscoveryMetaData): Moved components to components directory to improve maintainability
